+++ src/main/java/frc/robot/Constants.java
@@ -45,12 +45,12 @@ public final class Constants {
         public static final double slowSpeedMultiplier = 0.5;
         public static final double SlowTurnMultiplier = .3;
 
-        public static final COTSFalconSwerveConstants chosenModule =   //TODO: This must be tuned to specific robot Done!
+        public static final COTSFalconSwerveConstants chosenModule =
  COTSFalconSwerveConstants.SDSMK4i(COTSFalconSwerveConstants.driveGearRatios.SDSMK4i_L2);
 
         /* Drivetrain Constants */
-        public static final double trackWidth = Units.inchesToMeters(20.511811); //TODO: This must be tuned to specific robot Done!
-        public static final double wheelBase = Units.inchesToMeters(20.511811); //TODO: This must be tuned to specific robot Done!
+        public static final double trackWidth = Units.inchesToMeters(20.511811);
+        public static final double wheelBase = Units.inchesToMeters(20.511811);
         public static final double wheelCircumference = chosenModule.wheelCircumference;
 
         /* Swerve Kinematics 
@@ -95,22 +95,22 @@ public final class Constants {
         public static final double angleKF = chosenModule.angleKF;
 
         /* Drive Motor PID Values */
-        public static final double driveKP = 0.25; //TODO: This must be tuned to specific robot Done!
+        public static final double driveKP = 0.25;
         public static final double driveKI = 0.0;
         public static final double driveKD = 0.0;
         public static final double driveKF = 0.0;
 
         /* Drive Motor Characterization Values 
          * Divide SYSID values by 12 to convert from volts to percent output for CTRE */
-        public static final double driveKS = (0.45 / 12); //TODO: This must be tuned to specific robot Done!
+        public static final double driveKS = (0.45 / 12);
         public static final double driveKV = (1.85 / 12);
         public static final double driveKA = (0.000037994 / 12);
 
         /* Swerve Profiling Values */
         /** Meters per Second */
-        public static final double maxSpeed = 4.896; //TODO: This must be tuned to specific robot Done?
+        public static final double maxSpeed = 4.896;
         /** Radians per Second */
-        public static final double maxAngularVelocity = 5.25; //TODO: This must be tuned to specific robot: Done?
+        public static final double maxAngularVelocity = 5.25;
 
         /* Neutral Modes */
         public static final NeutralMode angleNeutralMode = NeutralMode.Coast;
@@ -118,7 +118,7 @@ public final class Constants {
 
         /* Module Specific Constants */
         /* Front Left Module - Module 0 */
-        public static final class Mod0 { //TODO: This must be tuned to specific robot Done!
+        public static final class Mod0 {
             public static final int driveMotorID = 6;
             public static final int angleMotorID = 8;
             public static final int canCoderID = 3;
@@ -128,7 +128,7 @@ public final class Constants {
         }
 
         /* Front Right Module - Module 1 */
-        public static final class Mod1 { //TODO: This must be tuned to specific robot Done!
+        public static final class Mod1 {
             public static final int driveMotorID = 7;
             public static final int angleMotorID = 2;
             public static final int canCoderID = 12;
@@ -158,7 +158,7 @@ public final class Constants {
         }
     }
 
-    public static final class AutoConstants { //TODO: The below constants are used in the example auto, and must be tuned to specific robot
+    public static final class AutoConstants {
         public static final double kMaxSpeedMetersPerSecond = 3;
         public static final double kMaxAccelerationMetersPerSecondSquared = 3;
         public static final double kMaxAngularSpeedRadiansPerSecond = Math.PI;
