--- src/main/java/frc/robot/commands/Autos/Test.java
@@ -1,36 +1,0 @@
-package frc.robot.commands.Autos;
-
-import edu.wpi.first.wpilibj2.command.InstantCommand;
-import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.commands.ArmToHome;
-import frc.robot.commands.IntakeCone;
-import frc.robot.commands.OutakeCone;
-import frc.robot.commands.OutakeCube;
-import frc.robot.commands.ScoreHigh;
-import frc.robot.commands.ScoreLow;
-import frc.robot.commands.ScoreMid;
-import frc.robot.commands.ShoulderToHome;
-import frc.robot.subsystems.ArmSubsystem;
-import frc.robot.subsystems.IntakeSubsystem;
-import frc.robot.subsystems.Swerve;
-import frc.robot.subsystems.WristSubsystem;
-//walker was here
-// NOTE:  Consider using this command inline, rather than writing a subclass.  For more
-// information, see:
-// https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class Test extends SequentialCommandGroup {
-  /** Creates a new Auto1. */
-  public Test(Swerve s_Swerve, IntakeSubsystem s_Intake, ArmSubsystem s_Arm, WristSubsystem s_Wrist) {
-    addCommands(
-
-    new PathPlannerCommand(s_Swerve, 2, "Back up after cone bottom", true)
-/*      
-.alongWith(new InstantCommand(() -> s_Arm.shoulderReversed *= -1))
-      .alongWith(new ScoreLow(s_Wrist, s_Arm))
-      .withTimeout(2),*/
-
-      
-
-    );
-  }
-}
