+++ src/main/java/frc/robot/commands/Drivetrain/Balance.java
@@ -4,8 +4,6 @@
 
 package frc.robot.commands.Drivetrain;
 
-import java.util.function.BooleanSupplier;
-
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.CommandBase;
