--- src/main/java/frc/robot/commands/TeleopSwerve.java
@@ -1,52 +1,0 @@
-package frc.robot.commands;
-
-import frc.robot.Constants;
-import frc.robot.subsystems.Swerve;
-
-import java.util.function.BooleanSupplier;
-import java.util.function.DoubleSupplier;
-
-import edu.wpi.first.math.MathUtil;
-import edu.wpi.first.math.geometry.Translation2d;
-import edu.wpi.first.wpilibj2.command.CommandBase;
-
-
-public class TeleopSwerve extends CommandBase {    
-    private Swerve s_Swerve;    
-    private DoubleSupplier translationSup;
-    private DoubleSupplier strafeSup;
-    private DoubleSupplier rotationSup;
-    private BooleanSupplier robotCentricSup;
-    private BooleanSupplier turnSlow;
-    private BooleanSupplier slow;
-
-    public TeleopSwerve(Swerve s_Swerve, DoubleSupplier translationSup, DoubleSupplier strafeSup, DoubleSupplier rotationSup, BooleanSupplier robotCentricSup, BooleanSupplier turnSlow, BooleanSupplier slow) {
-        this.s_Swerve = s_Swerve;
-        addRequirements(s_Swerve);
-
-        this.translationSup = translationSup;
-        this.strafeSup = strafeSup;
-        this.rotationSup = rotationSup;
-        this.robotCentricSup = robotCentricSup;
-        this.turnSlow = turnSlow;
-        this.slow = slow;
-    }
-
-    @Override
-    public void execute() {
-        /* Get Values, Deadband*/
-        double translationVal = MathUtil.applyDeadband(translationSup.getAsDouble(), Constants.stickDeadband);
-        double strafeVal = MathUtil.applyDeadband(strafeSup.getAsDouble(), Constants.stickDeadband);
-        double rotationVal = MathUtil.applyDeadband(rotationSup.getAsDouble(), Constants.stickDeadband);
-        if (this.turnSlow.getAsBoolean()) {rotationVal *= 0.3;}
-        if (this.slow.getAsBoolean()) {translationVal *= 0.2;}
-        if (this.slow.getAsBoolean()) {strafeVal *= 0.2;}
-        /* Drive */
-        s_Swerve.drive(
-            new Translation2d(translationVal, strafeVal).times(Constants.Swerve.maxSpeed), 
-            rotationVal * Constants.Swerve.maxAngularVelocity, 
-            !robotCentricSup.getAsBoolean(), 
-            true
-        );
-    }
-}
